Split webR unit tests into multiple files
