+++ src/tests/webR/utils.test.ts
@@ -1,0 +1,5 @@
+import { sleep } from '../../webR/utils';
+
+test('Utils sleep', async () => {
+  await expect(sleep(100)).resolves.not.toThrow();
+});
