+++ src/tests/webR/webr-main.test.ts
@@ -1,0 +1,73 @@
+import { WebR } from '../../webR/webr-main';
+import { Message } from '../../webR/chan/message';
+
+const webR = new WebR({
+  WEBR_URL: '../dist/',
+  RArgs: ['--quiet'],
+});
+
+jest.setTimeout(25000);
+
+describe('Test webR instance startup', () => {
+  test('Constructs successfully', () => {
+    expect(webR).toBeDefined();
+  });
+
+  test('Initialises successfully', async () => {
+    await expect(webR.init()).resolves.not.toThrow();
+  });
+});
+
+describe('Test webR simple console input/output', () => {
+  test('Wait for a prompt', async () => {
+    let msg: Message = await webR.read();
+    while (msg.type !== 'prompt') {
+      msg = await webR.read();
+    }
+    expect(msg.data).toBe('> ');
+  });
+
+  test('Write an R command to the console', () => {
+    expect(() => webR.writeConsole('42\n')).not.toThrow();
+  });
+
+  test('Read result line from stdout', async () => {
+    expect((await webR.read()).data).toBe('[1] 42');
+  });
+});
+
+describe('Evaluate R code', () => {
+  test('Evaluate R code without setting up error handlers', async () => {
+    const result = webR.evalRCode('webr::global_prompt_install()', undefined, {
+      withHandlers: false,
+    });
+    await expect(result).resolves.not.toThrow();
+  });
+
+  test('Throw an error if passed an invalid environment', async () => {
+    // @ts-expect-error Deliberate type error to test Error thrown
+    const promise = webR.evalRCode('3.14159', { env: 42 });
+    await expect(promise).rejects.toThrow('invalid environment object');
+  });
+
+  test('Throw an error if passed an invalid environment object type', async () => {
+    const euler = await webR.evalRCode('0.57722');
+    await expect(webR.evalRCode('x', euler)).rejects.toThrow('env argument with invalid SEXP type');
+  });
+
+  test('Throw errors from R', async () => {
+    const badSyntax = webR.evalRCode('42+');
+    await expect(badSyntax).rejects.toThrow('parse error');
+  });
+
+  test('Write to stderr while evaluating R code', async () => {
+    await webR.flush();
+    const res = webR.evalRCode('message("Hello, stderr!")');
+    await expect(res).resolves.not.toThrow();
+    expect((await webR.read()).data).toBe('Hello, stderr!');
+  });
+});
+
+afterAll(() => {
+  return webR.close();
+});
